Stop workflow commands

This will stop github from processing any workflow commands until octo_start_commands() is called with the correct token. This can be used if untrusted output (e.g. issue titles, bodies or commit messages) needs to be logged this can be used to stop this output from running possibly malicious workflow commands. Requires the suggested package openssl::openssl to be installed, when used within GitHub Actions.

octo_stop_commands()

octo_start_commands(token)

Arguments

token

A unique token used to restart workflow command parsing.

Value

The token needed to reactivate the workflow command parsing. When not on GitHub Actions returns a token not generated by openssl to avoid the dependency.

See also

Examples

Sys.setenv(GITHUB_ACTIONS = "true")
tk <- octo_stop_commands()
#> ::stop-commands::a97dd9b51f4c894f315f0385
# Commands will not be parsed by GitHub Actions
octo_start_commands(tk)
#> ::a97dd9b51f4c894f315f0385::